+++ CHANGELOG.md
@@ -6,6 +6,32 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.10] - 2020-08-30
+
+### Added
+
+* [NixOS/fonts] Add the option `confkit.fonts.enable` to enable a default
+    configuration for fonts. This includes additional fonts from which you can
+    opt-out by setting `confkit.fonts.installFonts` to `false`.
+* [Zsh/Git] Add `gi` for `git init` and `gsa` for `git submodule add`.
+* [Zsh/Git] Add more aliases for `hub` (see `zsh/git.zsh` for the full list).
+* [Zsh/Rust] Add `cic` to install from the current path.
+* [Zsh/Rust] Add `crr` for `cargo run --release`
+* [Zsh/Rust] Add `ce[r]` for `cargo embed [--release]`
+
+### Changed
+
+* [home-manager/screen] Install `screen` when enabling the module.
+* [Zsh/aliases] `e` Now opens the current directory with `codium` instead of
+    `code`.
+* [Zsh/FreeBSD] Remove Tmux exclusion from `pma`
+* [Zsh/Rust] Update `cs` and `csr` to the new `cargo-binutils` interface
+* [ranger/*] `E` now opens the current directory with `codium` instead of
+    `code`.
+* [Example] Reorganise the file layout and the sections inside the files.
+* [Example] Use Chrony to sync the time instead of the NTP reference
+    implementation.
+
 ## [0.0.9] - 2020-07-17
 
 ### Highlights
@@ -461,6 +487,7 @@ You have to update it to:
 
 * Extraction from my personal configuration framework.
 
+[0.0.10]: https://github.com/ejpcmac/confkit/compare/v0.0.9...v0.0.10
 [0.0.9]: https://github.com/ejpcmac/confkit/compare/v0.0.8...v0.0.9
 [0.0.8]: https://github.com/ejpcmac/confkit/compare/v0.0.7...v0.0.8
 [0.0.7]: https://github.com/ejpcmac/confkit/compare/v0.0.6...v0.0.7